Move tab update to context provider
